refactor: seperata component from chat component and useChat hook
